refactor: introduce lib.rs; refactor main.rs
